refactor: update Dockerfile for TCD repository cloning and reorganize ODM tests
- Changed the Dockerfile to clone the TCD repository as the deadtrees user instead of copying the source code.
- Deleted the outdated test_odm.py file and consolidated its content into a new test_process_odm.py file, improving test organization and maintainability.
- The new test_process_odm.py file includes comprehensive tests for ODM processing, covering execution, EXIF extraction, RTK detection, and orthomosaic generation.
